refactor(#1402): split suggest_actor_tree.py into nodes/suggest_actor/ subpackage
- suggest_actor_tree.py grew to 802 lines (threshold: 500, BTND-07-004/CS-18-002)
- Add nodes/suggest_actor/emit.py (401 lines): 4 emit node classes
- Add nodes/suggest_actor/conditions.py (168 lines): 3 duplicate-detection nodes
- Add nodes/suggest_actor/__init__.py: re-export facade
- Reduce suggest_actor_tree.py to 308 lines (factory functions only)
- Update case/nodes/__init__.py to re-export 7 new suggest_actor node classes
- All original import paths remain intact via backward-compat re-exports
- BTND07 structural ratchet: 68/68 passed; all 4834 unit tests pass
