+++ vultron/core/behaviors/case/suggest_actor_tree.py
@@ -28,15 +28,15 @@ in ADR-0026/CM-16:
 All three trees route through the CaseActor inbox per ADR-0021/ADR-0026 and
 use :func:`~vultron.core.behaviors.case.nodes.lifecycle.create_receive_activity_tree`
 to enforce CLP-10-006 ordering (ledger commit before effect nodes).
+
+BT leaf nodes for this workflow are in the
+:mod:`vultron.core.behaviors.case.nodes.suggest_actor` subpackage.
 """
 
 import logging
-from typing import cast
 
 import py_trees
-from py_trees.common import Status
 
-from vultron.core.behaviors.bridge import BTBridge
 from vultron.core.behaviors.case.nodes.actor import (
     EmitInviteActorToCaseNode,
     EvaluateDefaultRolesNode,
@@ -44,509 +44,17 @@ from vultron.core.behaviors.case.nodes.actor import (
 from vultron.core.behaviors.case.nodes.lifecycle import (
     create_receive_activity_tree,
 )
-from vultron.core.behaviors.helpers import DataLayerAction
-from vultron.core.behaviors.sync.commit_tree import (
-    create_commit_log_entry_tree,
+from vultron.core.behaviors.case.nodes.suggest_actor import (
+    ActorAlreadyParticipantNode,
+    EmitAcceptActorRecommendationNode,
+    EmitNoteDuplicateRecommendationToOwnerNode,
+    EmitOfferCaseParticipantToOwnerNode,
+    EmitRejectActorRecommendationNode,
+    InviteInFlightNode,
+    PendingOfferCaseParticipantNode,
 )
-from vultron.core.models.protocol_pair import (
-    INVITE_ACTOR_TO_CASE_REPLY_TYPES,
-    OFFER_CASE_PARTICIPANT_REPLY_TYPES,
-)
-from vultron.core.ports.case_persistence import CaseOutboxPersistence
-from vultron.enums.roles import CVDRole
 
 logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Shared emit nodes
-# ---------------------------------------------------------------------------
-
-
-class EmitOfferCaseParticipantToOwnerNode(DataLayerAction):
-    """Transform Offer(Actor, Case) → Offer(CaseParticipant) and DM Case Owner.
-
-    Uses ``trigger_activity_factory.offer_actor_to_case()`` with the
-    recommender_id, recommended_id, roles from the blackboard (written by
-    :class:`~vultron.core.behaviors.case.nodes.actor.EvaluateDefaultRolesNode`),
-    case_id, and case owner id.
-    The original offer ID is carried in the ``origin`` field so the Case Owner
-    can trace the causal chain (CM-16-004).
-
-    Reads ``suggested_roles_{recommendation_id_segment}`` from the blackboard
-    (namespaced per BTND-03-004); falls back to ``[CVDRole.VENDOR]`` when the
-    key is absent.
-
-    Returns SUCCESS on success, FAILURE if any required dependency is missing.
-    """
-
-    def __init__(
-        self,
-        recommendation_id: str,
-        recommender_id: str,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommendation_id = recommendation_id
-        self.recommender_id = recommender_id
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def setup(self, **kwargs) -> None:
-        super().setup(**kwargs)
-        id_segment = self.recommendation_id.split("/")[-1]
-        self.blackboard_key = f"suggested_roles_{id_segment}"
-        self.blackboard.register_key(
-            key=self.blackboard_key, access=py_trees.common.Access.READ
-        )
-
-    def _read_suggested_roles(self) -> list[CVDRole]:
-        try:
-            roles = self.blackboard.get(self.blackboard_key)
-            if isinstance(roles, list):
-                return roles
-        except KeyError:
-            pass
-        return [CVDRole.VENDOR]
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.feedback_message = "DataLayer or actor_id not available"
-            return Status.FAILURE
-
-        factory = self.trigger_activity_factory
-        if factory is None:
-            self.feedback_message = (
-                "trigger_activity_factory not in blackboard"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-        roles = self._read_suggested_roles()
-        if not roles:
-            self.feedback_message = (
-                f"suggested_roles for actor '{self.recommended_id}' is empty "
-                "— cannot emit Offer(CaseParticipant) without at least one role"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-        try:
-            case_obj = self.datalayer.read(self.case_id)
-            raw_owner = getattr(case_obj, "attributed_to", None)
-            owner_id = getattr(raw_owner, "id_", raw_owner) or None
-            if not owner_id:
-                self.feedback_message = (
-                    f"case '{self.case_id}' has no attributed_to (Case Owner) "
-                    "— cannot address Offer(CaseParticipant)"
-                )
-                self.logger.error(self.feedback_message)
-                return Status.FAILURE
-            activity_id, _ = factory.offer_actor_to_case(
-                recommender_id=self.recommender_id,
-                recommended_id=self.recommended_id,
-                case_id=self.case_id,
-                actor=self.actor_id,
-                origin=self.recommendation_id,
-                to=[owner_id],
-                roles=roles,
-            )
-            # Commit a local correlation marker first so duplicate checks work
-            # on retry even if the outbox write below fails (CM-16-008/AC-6).
-            # disposition="rejected" bypasses canonical-payload validation since
-            # Offer(CaseParticipant) is not a canonical ledger event type.
-            commit_tree = create_commit_log_entry_tree(
-                case_id=self.case_id,
-                object_id=self.recommended_id,
-                event_type="offer_case_participant",
-                disposition="rejected",
-            )
-            result = BTBridge(
-                datalayer=cast(CaseOutboxPersistence, self.datalayer)
-            ).execute_with_setup(
-                tree=commit_tree,
-                actor_id=self.actor_id,
-            )
-            if result.status != Status.SUCCESS:
-                raise RuntimeError(
-                    f"ledger correlation commit failed for "
-                    f"offer_case_participant/{self.recommended_id}"
-                )
-            cast(CaseOutboxPersistence, self.datalayer).record_outbox_item(
-                self.actor_id, activity_id
-            )
-            self.logger.info(
-                "%s: queued Offer(CaseParticipant) '%s' to Case Owner outbox"
-                " for case '%s'",
-                self.name,
-                activity_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        except Exception as e:
-            self.feedback_message = (
-                f"EmitOfferCaseParticipantToOwner failed: {e}"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-
-class EmitAcceptActorRecommendationNode(DataLayerAction):
-    """Queue AcceptActorRecommendation to the original recommender.
-
-    Used after the Case Owner accepts Offer(CaseParticipant) (CM-16-006 step 3).
-    The ``in_reply_to`` field is set to the original recommender's offer ID
-    (carried in the Case Owner's Accept via the ``origin`` field of the
-    transformed Offer) so the recommender can correlate the response.
-    """
-
-    def __init__(
-        self,
-        recommender_id: str,
-        recommendation_id: str,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommender_id = recommender_id
-        self.recommendation_id = recommendation_id
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.feedback_message = "DataLayer or actor_id not available"
-            return Status.FAILURE
-
-        factory = self.trigger_activity_factory
-        if factory is None:
-            self.feedback_message = (
-                "trigger_activity_factory not in blackboard"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-        try:
-            activity_id, _ = factory.emit_accept_actor_recommendation(
-                recommender_id=self.recommender_id,
-                recommendation_id=self.recommendation_id,
-                recommended_id=self.recommended_id,
-                case_id=self.case_id,
-                actor=self.actor_id,
-            )
-            cast(CaseOutboxPersistence, self.datalayer).record_outbox_item(
-                self.actor_id, activity_id
-            )
-            self.logger.info(
-                "%s: queued AcceptActorRecommendation to '%s' for case '%s'",
-                self.name,
-                self.recommender_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        except Exception as e:
-            self.feedback_message = (
-                f"EmitAcceptActorRecommendation failed: {e}"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-
-class EmitRejectActorRecommendationNode(DataLayerAction):
-    """Queue RejectActorRecommendation to the original recommender.
-
-    Used after the Case Owner rejects Offer(CaseParticipant) (CM-16-007 step 3).
-    """
-
-    def __init__(
-        self,
-        recommender_id: str,
-        recommendation_id: str,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommender_id = recommender_id
-        self.recommendation_id = recommendation_id
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.feedback_message = "DataLayer or actor_id not available"
-            return Status.FAILURE
-
-        factory = self.trigger_activity_factory
-        if factory is None:
-            self.feedback_message = (
-                "trigger_activity_factory not in blackboard"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-        try:
-            activity_id, _ = factory.emit_reject_actor_recommendation(
-                recommender_id=self.recommender_id,
-                recommendation_id=self.recommendation_id,
-                recommended_id=self.recommended_id,
-                case_id=self.case_id,
-                actor=self.actor_id,
-            )
-            cast(CaseOutboxPersistence, self.datalayer).record_outbox_item(
-                self.actor_id, activity_id
-            )
-            self.logger.info(
-                "%s: queued RejectActorRecommendation to '%s' for case '%s'",
-                self.name,
-                self.recommender_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        except Exception as e:
-            self.feedback_message = (
-                f"EmitRejectActorRecommendation failed: {e}"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-
-# ---------------------------------------------------------------------------
-# Duplicate-detection precondition nodes (CM-16-008, CM-16-009)
-# ---------------------------------------------------------------------------
-
-
-class ActorAlreadyParticipantNode(DataLayerAction):
-    """Return SUCCESS if the recommended actor is already a case participant.
-
-    Reads ``VulnerabilityCase.actor_participant_index`` from the DataLayer.
-    Returns SUCCESS when ``recommended_id`` is a key in the index (AC-7b,
-    CM-16-009).
-
-    Used as the first arm of the duplicate-detection Selector in
-    :func:`create_recommend_actor_to_case_received_tree`.
-    """
-
-    def __init__(
-        self,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.feedback_message = "DataLayer not available"
-            return Status.FAILURE
-        case_obj = self.datalayer.read(self.case_id)
-        index = getattr(case_obj, "actor_participant_index", {}) or {}
-        if self.recommended_id in index:
-            self.logger.info(
-                "%s: actor '%s' is already a participant in case '%s'",
-                self.name,
-                self.recommended_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        return Status.FAILURE
-
-
-class InviteInFlightNode(DataLayerAction):
-    """Return SUCCESS if an Invite to the recommended actor is in-flight.
-
-    Queries the case ledger via ``find_protocol_pair`` with
-    ``event_type="invite_actor_to_case"`` and ``object_id=recommended_id``.
-    Returns SUCCESS when the pair ``is_pending()`` — i.e., an Invite was
-    sent and no Accept/Reject has been recorded (AC-7a, CM-16-009).
-
-    Used as the second arm of the duplicate-detection Selector in
-    :func:`create_recommend_actor_to_case_received_tree`.
-    """
-
-    def __init__(
-        self,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.feedback_message = "DataLayer not available"
-            return Status.FAILURE
-        pair = self.datalayer.find_protocol_pair(
-            case_id=self.case_id,
-            request_event_type="invite_actor_to_case",
-            object_id=self.recommended_id,
-            reply_event_types=INVITE_ACTOR_TO_CASE_REPLY_TYPES,
-        )
-        if pair.is_pending():
-            self.logger.info(
-                "%s: Invite to '%s' is in-flight for case '%s'",
-                self.name,
-                self.recommended_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        return Status.FAILURE
-
-
-class PendingOfferCaseParticipantNode(DataLayerAction):
-    """Return SUCCESS if an Offer(CaseParticipant) to the Case Owner is pending.
-
-    Queries the case ledger via ``find_protocol_pair`` with
-    ``event_type="offer_case_participant"`` and ``object_id=recommended_id``.
-    Returns SUCCESS when the pair ``is_pending()`` — i.e., the CaseActor
-    has already forwarded an Offer(CaseParticipant) for this actor and the
-    Case Owner has not yet responded (AC-6, CM-16-008).
-
-    Used as the third arm of the duplicate-detection Selector in
-    :func:`create_recommend_actor_to_case_received_tree`.
-    """
-
-    def __init__(
-        self,
-        recommended_id: str,
-        case_id: str,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.feedback_message = "DataLayer not available"
-            return Status.FAILURE
-        pair = self.datalayer.find_protocol_pair(
-            case_id=self.case_id,
-            request_event_type="offer_case_participant",
-            object_id=self.recommended_id,
-            reply_event_types=OFFER_CASE_PARTICIPANT_REPLY_TYPES,
-        )
-        if pair.is_pending():
-            self.logger.info(
-                "%s: Offer(CaseParticipant) for '%s' is pending Case Owner "
-                "decision in case '%s'",
-                self.name,
-                self.recommended_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        return Status.FAILURE
-
-
-class EmitNoteDuplicateRecommendationToOwnerNode(DataLayerAction):
-    """Send a Note DM to the Case Owner noting reinforcing demand.
-
-    Used when a second ``Offer(Actor, Case)`` arrives while a first
-    ``Offer(CaseParticipant)`` is still pending the Case Owner's
-    Accept/Reject decision (AC-6, CM-16-008).  Sends a
-    ``Create(Note)`` + ``Add(Note, Case)`` to the Case Owner without
-    issuing a second ``Offer(CaseParticipant)``.
-
-    When ``offer_content`` is provided (non-None, non-empty), it is appended
-    to the Note body per CM-16-008's requirement to forward the incoming
-    Offer's ``content`` field to the Case Owner.
-    """
-
-    def __init__(
-        self,
-        recommendation_id: str,
-        recommender_id: str,
-        recommended_id: str,
-        case_id: str,
-        offer_content: str | None = None,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.recommendation_id = recommendation_id
-        self.recommender_id = recommender_id
-        self.recommended_id = recommended_id
-        self.case_id = case_id
-        self.offer_content = (
-            (offer_content.strip() or None) if offer_content else None
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.feedback_message = "DataLayer or actor_id not available"
-            return Status.FAILURE
-
-        factory = self.trigger_activity_factory
-        if factory is None:
-            self.feedback_message = (
-                "trigger_activity_factory not in blackboard"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-        try:
-            case_obj = self.datalayer.read(self.case_id)
-            raw_owner = getattr(case_obj, "attributed_to", None)
-            owner_id = getattr(raw_owner, "id_", raw_owner) or None
-            if not owner_id:
-                self.feedback_message = (
-                    f"case '{self.case_id}' has no attributed_to (Case Owner) "
-                    "— cannot address duplicate-recommendation Note"
-                )
-                self.logger.error(self.feedback_message)
-                return Status.FAILURE
-
-            actor_segment = self.recommended_id.split("/")[-1]
-            content = (
-                f"Duplicate actor recommendation: '{self.recommender_id}' "
-                f"also recommends actor '{actor_segment}' for case "
-                f"'{self.case_id}'.  An Offer(CaseParticipant) is already "
-                f"pending your decision (no second offer sent)."
-            )
-            if self.offer_content:
-                content = (
-                    f"{content}\n\nRecommender note: {self.offer_content}"
-                )
-            note_id, _ = factory.create_note(
-                name=f"Duplicate recommendation — {actor_segment}",
-                content=content,
-                context_id=self.case_id,
-                attributed_to=self.actor_id,
-                in_reply_to=self.recommendation_id,
-            )
-            activity_id, _ = factory.add_note_to_case(
-                note_id=note_id,
-                case_id=self.case_id,
-                actor=self.actor_id,
-                to=[owner_id],
-            )
-            cast(CaseOutboxPersistence, self.datalayer).record_outbox_item(
-                self.actor_id, activity_id
-            )
-            self.logger.info(
-                "%s: sent duplicate-recommendation Note to Case Owner '%s'"
-                " for case '%s'",
-                self.name,
-                owner_id,
-                self.case_id,
-            )
-            return Status.SUCCESS
-        except Exception as e:
-            self.feedback_message = (
-                f"EmitNoteDuplicateRecommendationToOwner failed: {e}"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-
-# ---------------------------------------------------------------------------
-# BT factory functions
-# ---------------------------------------------------------------------------
 
 
 def create_receive_offer_case_participant_tree(
@@ -788,6 +296,7 @@ def create_reject_actor_recommendation_received_tree(
 
 
 __all__ = [
+    # node classes live in nodes.suggest_actor; re-exported here for backward compat
     "ActorAlreadyParticipantNode",
     "EmitAcceptActorRecommendationNode",
     "EmitNoteDuplicateRecommendationToOwnerNode",
